vm: convert runtime errors from internal to external representation
The "append: out of memory" error was returned as "reflect.Append:
slice overflow".

The "makechan: size out of range" was returned as "reflect.MakeChan:
negative buffer size".

The "close of nil channel" and "send on closed channel" errors were
returned as a "runtime" package error and not as a "vm" package error.

Update #292