compiler: fix panic in function 'emitTemplate'
If the code inside 'emitTemplate' panics and the error is not a
*LimitExceededError then a nil panic is raised.
This is not the indented behavior; this commit fixes the deferred
function inside emitTemplate that references the variable 'err' when
it has not been set.

This commit is similar to the commit 26fdb8ffe9e315dafc739b680dc537187ffeb41a.