compiler/emitter: disable imports with init functions in templates
This commit disables (by panicking) the 'import' statements in templates
that refer to template pages with 'init' functions. That's because the
current implementation contains an invalid behavior: the init functions
are called many times as the page is imported. The test ""Init function
called more than once", for example, panics if it is executed.

Also disables some tests and add some commented tests, waiting for #643
to be resolved.

For more information please see the issue #643.