compiler/emitter: rename a field of the type 'emitter'
This commit renames the field 'alreadyInitializedPkgs' of the type
'emitter' to 'alreadyInitializedTemplatePkgs'.

That's because the old name was wrong: the imported packages for Go
programs (i.e. not templates) are not tracked in this map. The field
'alreadyInitializedPkgs' keeps track only of the imported template
packages.

The initialization of imported Go packages is handled in a different
manner by the emitter, that is not related to this field.