compiler: fix error message when a template comment is not terminated
In templates, if a comment is not terminated the error message is:

unexpected EOF, expecting #}

with position at the start of the comment and not at the end of the
file as would expect from the error message.

But if the position was at the end of the file, it would not be possible
to know where the unclosed comment begins.

So this commit changes the error message with the same message for
an unclosed Go comments:

comment not terminated