compiler: change handling of tags context in lexer
This commit does not change the behaviour of Scriggo but changes the
type "lexer" and the checks that change the lexer context from HTML to
JavaScript or CSS during the lexing.

This commit is preparatory to the next commit that makes more strict the
lexing of script and style tags.