compiler: make lexing of script and style tags more compliant
A browser executes a script only if its tag does not have the attribute
"type" or if its value is empty or if its value is "text/javascript"
in a case insensitive form with or without leading and trailing spaces.

The same is true for style tags but with the "text/css" value.

This commit changes the lexer to follow these browser behaviours.