compiler: fix error message returned on cycle in templates
This commit fixes the error message returned if there is a cycle parsing
a template file. So instead of:

cycle not allowed
/index.html
	imports imports include imports /includes/macro.html

the following error message is returned:

file /index.html
	extends  /layout.html
	imports  /includes/macro.html
	includes /includes/include.html
	imports  /includes/macro.html: cycle not allowed

It also changes the error message on cycles in programs moving the text
"import cycle not allowed" at the end of the message as gc does.