compiler/checker: fix method fieldByName
When executing this code:

    package main

    import "time"

    func main() {
        _ = time.Now().loc
    }

the field 'loc' is not accessible. That's not because Scriggo checks
that 'loc' is not exported, but this is caused by a bug in the method
'fieldByName'; such method calls 'IsPredefined' on a type info, but the
information about where the type has been defined is lost and the value
returned by 'IsPredefined' is not relevant in this context.

So, this commit fixes the method 'fieldByName'. Doing so, the code above
is executed as if it is valid and that's an invalid behavior. So,
another commit will follow this one that will fix such invalid behavior.