compiler: store type infos into 'compilation' instead of 'checker'
Take a look at the issue https://github.com/open2b/scriggo/issues/661
for more background about this commit.

This commit fixes the issue #661 but introduces a bug in the system that
checks if a package has been imported but not used.

We decided to give priority to the issue #661, so the test
'test/compare/testdata/program_imports/many_calls_to_init.go' has been
skipped because it reports a false positive 'imported but not used'.

Fix #661.