compiler/checker: fix sorting of packages imported multiple times
When a package is imported twice (for example in templates) the type
checker panics when it tries to sort the declarations.

This commit fixes that behavior by sorting the declarations only the
first time that the type checker encounters a package.

See the issue #668 for more details about this commit.

Fix #668.