compiler/emitter: fix shadowing of global vars with ':='
In the following scenario:

* a global variable is shadowed by a local identifier in a ':='
  assignment
* the global variable is referenced in the expression on the right side
  of the ':=' assignment

The runtime panics because the emitter emits bad code; that happens
because when the emitter encounters the global variable on the rhs it
handles it as the local variable instead of the global variable.

This commit fixes that bad behavior by mapping the name of the local
variables to the corresponding register only when all the expressions on
the rhs of the ':=' assignment have been emitted.

Fix #679.