compiler/emitter: fix stack shift when emitting initialization funcs
When emitting the initialization functions for imported packages the
stack shift is [0,0,0,0], so some values in registers may be erroneously
overwritten by the parameters of initialization functions.

(Note: 'initialization functions' include both 'init' functions as well
as dummy functions that initializes the variables of the imported file).

This commit fixes the stack shift.

Fix #672.