all: add typeof parameter to the (*vm).Run and implement in the compiler
This is part of a series of commits that fix the If instruction when
comparing two interface values when one or both have a Scriggo type.

This commit adds a new function parameter to the (*vm).Run method that
it is intended to return the reflect type of a reflect value where the
value can be a wrapped Scriggo type value.

A following commit will fix the If instruction, in the emitter and
runtime, using the typeof function.