compiler/parser: fix early loop termination parsing 'import' statement
If a 'import' statement is preceded by an 'extends' or an 'import'
statement, it is not necessary to continue to check the previous sibling
nodes. But, in current code, the 'break' statement terminate the
'switch' and not the 'for' as it was intended to do.

This is not a parser invalid behaviour but it is a buggy code that
does not optimize what was intended to optimize. This commit fixes it.