compiler/checker: fix checking of partial files
Currently, when a partial file is checked, the top level declarations in
that file are added to the file/package block of the shown file.

This is wrong because a shown file should only have local scopes, and
that leads to runtime errors with code that contains upvars.

This commit fixes that bug by creating a local scope that embraces all
the shown file.

Fix #694.