compiler/parser: use term 'imported' instead of 'declarationsFile'
'declarationsFile' is used as name for some variables and parameters to
indicates whether a parsed file is a declaration file but in practise
these variables and arguments are true only when the file is imported.

Since the term 'imported' it is more clear that 'declarationsFile', this
change renames these variables and parameters to 'imported'.