compiler/parser: improve error messages when text in declaration files
This change improves the error message when a not empty text is used in
a declaration file. Before this change the error is 'unexpected text,
expecting declaration', but it is not clear why a declaration is
expected and this is inaccurate because not only declarations are
allowed.

This commit changes the error messages to 'unexpected text in imported
file' and 'unexpected text in file with extends'.