templates: get file language from file extension by default
This commit changes the Load method to read the file's language from the
file's extension by default. If the provided FileReader implements the
method 'Language(string) (Language, error)', this method is called to
get the language of the file.

Consequently the compiler.FileReader.ReadFile method is changed to
return the language of the file.