all: replace TypeOf function with the Types interface
The runtime uses the reflect.New function also for Scriggo types, but
the function panics because a Scriggo reflect.Type cannot be used with
the reflect functions.

This commit replaces the TypeOf function with a Types interface that
exposes, other than the TypeOf method, the methods ValueOf and New.

Next commits will fix the uses of the runtime.New function in the
runtime.