compiler,runtime: disassemble macros as 'Macro' instead of 'Func'
This change disassembles macros using a new assembly instruction 'Macro'
instead of 'Func' and using 'macro' as an alias for 'func'.

For this, it adds the 'Macro' field to the 'runtime.Function' type to
indicates if it is a macro. It also updates the compiler to set this
field for macros.

It also disassembles all the functions and macros in a template instead
of disassembling only the main function.