all: allow Markdown to extends HTML and HTML to show imported Markdown
This change allows a Markdown file to extends an HTML file and allows an
HTML file to show a macro of an imported Markdown file.

This change does not allow yet to show a Markdown partial from an HTML
file.

The 'Run' method of '*templates.Template' accepts the new
'MarkdownConverter' option that provides the function that converts
Markdown to HTML.

For #432