test/compare: fix 'compare' command for Go 1.16
The following error occurs executing the 'compare' command with Go 1.16:

    main.go:7:2: module testpkg provides package testpkg and is replaced
    but not required; try 'go get -d testpkg' to add it

The error does not occur anymore adding

    require testpkg v0.0.0-00010101000000-000000000000

to the go.mod file in the temporary module directory. But a second error
occurs:

    go: testpkg@v0.0.0-00010101000000-000000000000 requires
        github.com/open2b/scriggo@v0.0.0 requires
        github.com/google/go-cmp@v0.5.1: missing go.sum entry; try
        'go mod download github.com/google/go-cmp' to add it

Adding a 'go.sum' file in the the temporary module directory the error
does not occur anymore and the 'compare' command runs.