all: replace the FileReader interface with the fs.FS interface
This change replaces the templates.FileReader interface with the fs.FS
interface introduced in Go 1.16 and uses the fs syntax for file paths.

It also changes the signature for templates.Build,
compiler.BuildTemplate and compiler.ParseTemplate methods and introduces
the templates.FormatFS (and compiler.FormatFS) interface to read a file
format.