all: replaces 'show' statements with a single 'show' statement
This change replaces the two statements 'show(<expr>)' and
'show <macro>' with the statement 'show <expr>' with the same behaviour
of 'show(<expr>)'.

It also

- does not allow the 'show' statement in functions but only in macros
- fixes the fromFormat in macro calls to the real format and not the
  page format
- renames the 'Call' operation to 'CallFunc' and adds the 'CallMacro'
  operation
- changes the runtime.Renderer renaming 'End' to 'Close' and splitting
  the 'Enter' method in 'WithOut' and 'WithConversion' methods.

Note that the statement 'show <path>' continues to be deprecated and to
show a string literal with the 'show' statement you have to put the
string in brackets.

For #698