compiler/parser: change error message if 'goto' is outside function body
The error message, in templates, if the 'goto' statement is outside
function body is "unexpected goto outside function body" but for the
'return' statement is "return statement outside function body".

This change makes the error for 'goto' like the error for 'return' and
also rewrites the condition that checks the error to use 'end' instead
of the token's context.