compiler/emitter: optimize Text instructions merging them into one
This change merges more Text instructions in a single instruction when
possible. Tested on a simple HTML template, the execution time was cut
in half.

The emitter can emit consecutive Text instructions after a tree
transformation and if there are URL nodes with a single Text node.

For the following code

  <a href="https://en.wikipedia.org/">Wikipedia</a>

this change emits

  Text "<a href=\"https://en.wikipedia.org/\">Wikipedia</a>"

instead of

  Text "<a href=\""
  Text "https://en.wikipedia.org/"
  Text "\">Wikipedia</a>"

Other cases of merge are possible, but these two are the most common.