templates: move MarkdownConverter option from run to build
This change moves the MarkdownConverter option from the options of the
'Run' method to the options of the 'Build' method. This is done in
preparation to the change that adds the explicit conversion from a
markdown value to an html value.

This change also makes the '(*.Template).Run' method to return an error
if 'out' is nil. io.discard can be used instead of nil.