runtime, compiler: simplify Move instruction removing explicit 'copy'
The 'emitMove' function of the function builder takes a 'copy' argument
that should be true only if the source register contains an array or a
struct.

But only one call to emitMove (the last call in the '_changeRegister'
method of the emitter) can fall into this case and for the other calls
the 'copy' argument can be false.

Apart from this, the 'copy' parameter is redundant because the 'Move'
instruction already checks if the source register contains an array or
a slice.

This change then removes the 'copy' parameter from the 'emitMove'
function and joins the two 'generalRegister' and '-generalRegister'
cases, in the 'Move' instruction, into one case.