runtime, compiler: use new 'Function.FieldIndexes' for field indexes
The field indexes are currently stored encoded as int64 in the
'Function.Constants.Ints' slice. This limits the number of integer
constants and field indexes to 256, and limits the number of fields per
struct to 256.

This change stores the field indexes, no longer encoded as int64, in the
new 'Function.FieldIndexes' slice. In this way, the limits of the
constants can be increased in the future, the number of struct fields
are no longer limited to 256 and accessing the struct fields does not
require decoding from int64.