compiler/parser: simplify and fix function parameters parsing
The code that parses the function parameters is difficult to follow and
contains some errors.

This declaration is valid

  func f11(a int, b ...string,) {}

but the parser fails with error "can only use ... with final parameter
in list".

This declaration

  func f(*int, string)

should fail with error "mixed named and unnamed function parameters" but
fails with error "unexpected *int, expecting )".

Fixes #541