all: use 'native' and 'non-native' instead of 'predefined' and 'Scriggo'
This change uses the 'native' term to refer to packages, types,
variables, functions and values pre-compiled together with the Scriggo
compiler and runtime in an executable file. It uses 'non-native' to
refer to packages, types, variables, functions and values in a source
code compiled by Scriggo.

It also use 'source code', instead of 'Scriggo code', to refer to a
source code that is compiled by Scriggo.

To do this it changes comments, documentation and renames types,
functions, variables, constants and fields to use the new terms and
renames the 'CallPredefined' instruction to 'CallNative'.

Closes #381