compiler/emitter: panic in case of indirect of for range variable
This commit adds some code that checks if a for range variable should be
declared as indirect; in that case it panics in the compiler instead of
panicking in the runtime returning an obscure error message after that
some operations may be already executed.

See the issue https://github.com/open2b/scriggo/issues/629 for more
information about this.