test/compare: don't call 'mustBeAssignableTo' checking 'x op= y'
The call to 'mustBeAssignableTo' cannot fail because 'x op= y' is
equivalent to 'x = x op (y)' and the type checker has already checked
that 'x' can be used as left value and that 'x op (y)' is valid. Since
the type of 'x op (y)' is identical to the type of 'x', 'x op (y)' is
assignable to 'x'.

The '<<=' and '>>=' assignments have no a special behavior. The special
cases of the shift operator do not apply because 'x' is not a constant.