scriggo: remove 'IsLimitExceeded' function
The compiler can return four type of errors: SyntaxError, CheckingError,
LimitExceededError and CycleError. The scriggo package has the function
'IsLimitExceeded' to check if an error is a limit execeeded error but
has no similar functions for the other types.

This change removes the 'IsLimitExceeded' instead of adding tree more
functions.