compiler/checker: fix upvars for predefined values in templates
This commit removes the check 'nestedFuncs[0].Ident == nil' and
simplifies the code.

Such check was probably added with a previous implementation of the
macros; currently it creates bad upvars for functions and macros nested
into macros at package-level in templates.

See the issue #768 for more information about this commit.

Fix #768.