compiler/checker: fix unreported errors in type switch
This commit fixes some unreported errors in the type switch. For the
following example, the type checker does not report the error
'x declared but not used'

  switch x := interface{}(0).(type) { case int: }

For the following example, the type checker does not report the error
'x redeclared in this block'

  switch x := interface{}(0).(type) { case int: var x = 5; _ = x }

This commit fixes these issues and reports the errors in the same
position reported by gc.

Fixes #474