compiler/checker: fix imported and not used error enabling package cache
This change enable the package cache that was disable by commit
6763a0d923e3b47b52b69a5274c887f432f9ddfb.

In this way it fixes the issue #662 about the "imported and not used"
error which occurs when a package is imported two or more times.

The issue #662 is due to multiple type checks of packages. In the second
type check the identifiers in a function body are not checked because
they have already been checked in the first type check and consequently
they have not been set as used in scopes.

Fixes #662