compiler/lexer: fix parsing BOM as first code point of the file
If the program file begins with the BOM code point, a syntax error
occurs because the lexer skips only the first byte of the encoding of
the code point.

This change fixes also the tests passing the file unchanged to Scriggo.

Furthermore, if there had not been this issue, the column number of any
subsequent error in the first line would have been incorrect.

Refers #803