test/compare/cmd: add a workaround for 'go mod tidy'
This commit explicitly imports the package "testpkg". Otherwise, if the
file 'predefPkgs.go' has not been populated by 'go generate', running
'go mod tidy' cleans up the 'go.mod' removing the dependency from the
module 'testpkg' (as it is not used elsewhere in this module) and the
next 'go generate' fails.