cmd/scriggo: fix import of the package 'reflect'
Currently, when trying to compile a package that does not refer to the
'reflect' package in its declarations, the build of the interpreter
fails because the 'reflect' package has been imported but not used.

Due to this issue, currently it's not possibile to build and interpreter
that contains, for example, only the "math" package.

This commit fixes it.