compiler/checker: differentiate error messages in type declaration loops
Currently, the type checker reports the same error both when
encountering a type alias declaration loop and a type definition loop.

This commit refactors the function 'compiler.detectTypeLoop' and fixes
that behavior.

Although this commit is not strictly related to the issue #440 it may
fix some cases, making the solving of that issue easier.