compiler/types: fix creation of defined types
Currently, when two defined types are declared using the package
'compiler/types', if their name and their underlying type is the same
then the resulting 'reflect.Type's are the same.

This violates the rules for which two 'reflect.Type's can be equal if
and only if the types that they represent are identical.

This commit fixes that by creating different reflect.Types for different
defined types, that are always not-identical from other types.

Fix #810.