compiler/checker: fix checking of struct fields
The type checker should fail with error 'embedded type cannot be a
pointer to interface' if a field has a pointer to interface type but it
panics. For example, type checking the following invalid struct types,
it panics

  type T interface{}; _ = struct{ T }{}
  type T interface{}; _ = struct{ *T }{}
  type T *interface{}; _ = struct{ T }{}

The type checker does not check for duplicated fields between embedded
fields. As a consequence, it panics checking the following illegal
struct type

  type T int; _ = struct{ T int; T }{}

The type checker panics with the following illegal struct types

  type T *int; _ = struct{ *T }{}
  type T *int; type A = T; _ = struct{ *A }{}

The type checker should fail with error 'embedded type cannot be a
pointer' for the following illegal struct types, but it does not fail

  type T *int; _ = struct{ T }{}
  type T *int; type A = T; _ = struct{ A }{}

If an embedded field is a type alias, the name of the field should be
the alias but the type checker uses the name in the type definition. As
a consequence the following valid code fails

  type T int; type A = T; _ = struct{ A }{}.A

This change fixes all these issues.