compiler/checker: fix corner cases in field selector
This change implements field finding by name instead of using the
reflect method of the struct type. This resolve some corner cases in
the find.

For example the 'fieldByName' method checks if the field is exported
before checking if there is an ambiguous selector error to return.

Checking if the field can be accessed by the code in the current
package is done on the main struct instead of the struct where the field
is located.

This change fixes these issues.