cmd/scriggo: add an initial implementation of 'scriggo bug'
The sub-command 'go bug' opens a web browser with a partially filled
form that can be used to report a bug.

In Scriggo we also added the sub-command 'scriggo bug', that was
unimplemented until now.

This commit adds an initial implementation of such sub-command.

Note that opening a web browser, just like 'go bug' does, requires a lot
of code to handle different system configuration/different
oses/different web browsers, so it has not been implemented for now.