compiler/checker: fix call to 'setValue'
Currently, when assigning a value of a type T1 to a variable of type T2,
if such assignment requires an implicit conversion then such conversion
is performed at runtime, calling the 'OpConvert'.

This is not correct and leads to problems.

This commit avoids that by calling 'setValue' passing it the correct
type instead of nil, so the conversion is performed at compile time.

Fix #917.

Co-authored-by: Marco Gazerro <gazerro@open2b.com>