compiler/checker: don't consider 'make(...)' a constant function call
Go version 1.18 does not consider 'make (...)' to be a constant function
call when used as argument expression of 'len' and 'cap'. Therefore
these assignments

const c1 = len(make([][1]int, 1)[0])
const c1 = cap(make([][1]int, 1)[0])

are valid in Go 1.17 and earlier and not valid in Go 1.18.

This change makes them invalid in Scriggo.