test/misc: declare multi file test cases within testing function
This is convenient because many editors allow you to run the test
function directly where the cursor is located. Currently, however, since
the tests are declared outside the test function, this requires for
navigating up and down the file to run the tests when one of them is
modified.

For this reason, this commit declares multi-file test cases within the
testing function.